api/pyproject.toml
[tool.poetry]
name = "base"
version = "0.1.0"
description = "Python base project"
authors = ["son@zen8labs.com"]

[tool.poetry.dependencies]
Django = "^4.2.3"
django-filter = "^21.1"
django-phonenumber-field = "^6.4.0"
django-rest-auth = "^0.9.5"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
phonenumbers = "^8.13.16"
Pillow = "^9.5.0"
psycopg2-binary = "^2.9.6"
python = "^3.11"
requests = "^2.31.0"
argon2-cffi = "^21.3.0"
uvicorn = "^0.18.3"
django-admin-rangefilter = "^0.10.0"
boto3 = "^1.28.15"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
